create activities, moved schemas, methods, publications, update imports || moved lib folder no longer needed
