+++ imports/ui/RouterComponents/activities/ActivitiesContainer.js
@@ -1,6 +1,8 @@
 import { withTracker } from 'meteor/react-meteor-data';
 import ActivitiesList from './Activities';
+
 import Processes from '../../../api/processes/process';
+import Activities from '../../../api/activities/activity';  
 
 export default ActivitiesContainer = withTracker((props) => {
   const processesSubscription = Meteor.subscribe('processes');
