--- lib/collections.js
@@ -1,13 +1,0 @@
-import { Mongo } from 'meteor/mongo';
-
-// Hosts = new Mongo.Collection('hosts');
-
-Activities = new Mongo.Collection('activities');
-// Processes = new Mongo.Collection('processes');
-// Works = new Mongo.Collection('works');
-// Pages = new Mongo.Collection('pages');
-
-// Resources = new Mongo.Collection('resources');
-// Chats = new Mongo.Collection('chats');
-// Documents = new Mongo.Collection('documents');
-// Categories = new Mongo.Collection('categories');
